Trained model. 40 Seq Len, 20 epochs.
This submission contains the checkpoint for the LSTM model trained over 20 epochs on the dataset. No dropout or other regularization techniques were used, however the optimizer was changed to RMS with a learning rate of .01. Additionally, some of the code cells were split up into smaller chunks so as to minimize the amount of code that was re-rerun each time.